+++ packages/editor/src/private-apis.js
@@ -8,7 +8,6 @@ import * as interfaceApis from '@wordpress/interface';
  */
 import { lock } from './lock-unlock';
 import { EntitiesSavedStatesExtensible } from './components/entities-saved-states';
-import EditorContentSlotFill from './components/editor-interface/content-slot-fill';
 import BackButton from './components/header/back-button';
 import Editor from './components/editor';
 import PluginPostExcerpt from './components/post-excerpt/plugin';
@@ -26,6 +25,10 @@ import {
 } from '@wordpress/fields';
 import { registerCoreBlockBindingsSources } from './bindings/api';
 import { getTemplateInfo } from './utils/get-template-info';
+import GlobalStylesUIWrapper from './components/global-styles';
+import { StyleBookPreview } from './components/style-book';
+import { useGlobalStyles, useStyle } from './components/global-styles/hooks';
+import { GlobalStylesActionMenu } from './components/global-styles/menu';
 
 const { store: interfaceStore, ...remainingInterfaceApis } = interfaceApis;
 
@@ -37,7 +40,6 @@ lock( privateApis, {
 	BackButton,
 	EntitiesSavedStatesExtensible,
 	Editor,
-	EditorContentSlotFill,
 	PluginPostExcerpt,
 	PostCardPanel,
 	PreferencesModal,
@@ -48,6 +50,12 @@ lock( privateApis, {
 	ResizableEditor,
 	registerCoreBlockBindingsSources,
 	getTemplateInfo,
+	// Global Styles
+	GlobalStylesUIWrapper,
+	GlobalStylesActionMenu,
+	StyleBookPreview,
+	useGlobalStyles,
+	useStyle,
 	// This is a temporary private API while we're updating the site editor to use EditorProvider.
 	interfaceStore,
 	...remainingInterfaceApis,